+++ .devcontainer/hooks/updateContent.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 set -eux
 
+WORKSPACE_DIR=$(realpath /workspaces/*)
+
 npm install
 npm install --prefix "$HOME/twinejs"
 poetry install -P "$WORKSPACE_DIR/spindler"